refactor aggregator to prepare it for feijoa changes (split aggregator.go file and rename some functions)
