+++ aggregator/aggregate_blobouter.go
@@ -1,0 +1,7 @@
+package aggregator
+
+import "context"
+
+func (a *Aggregator) tryAggregateBlobOuterProofs(ctx context.Context, prover proverInterface) (bool, error) {
+	return false, nil
+}
