+++ aggregator/batch.go
@@ -1,0 +1,194 @@
+package aggregator
+
+import (
+	"context"
+	"encoding/json"
+	"errors"
+	"fmt"
+	"math/big"
+	"time"
+
+	"github.com/0xPolygonHermez/zkevm-node/log"
+	"github.com/0xPolygonHermez/zkevm-node/state"
+)
+
+func (a *Aggregator) tryGenerateBatchProof(ctx context.Context, prover proverInterface) (bool, error) {
+	log := log.WithFields(
+		"prover", prover.Name(),
+		"proverId", prover.ID(),
+		"proverAddr", prover.Addr(),
+	)
+	log.Debug("tryGenerateBatchProof start")
+
+	batchToProve, proof, err0 := a.getAndLockBatchToProve(ctx, prover)
+	if errors.Is(err0, state.ErrNotFound) {
+		// nothing to proof, swallow the error
+		log.Debug("Nothing to generate proof")
+		return false, nil
+	}
+	if err0 != nil {
+		return false, err0
+	}
+
+	log = log.WithFields("batch", batchToProve.BatchNumber)
+
+	var (
+		genProofID *string
+		err        error
+	)
+
+	defer func() {
+		if err != nil {
+			err2 := a.State.DeleteBatchProofs(a.ctx, proof.BatchNumber, proof.BatchNumberFinal, nil)
+			if err2 != nil {
+				log.Errorf("Failed to delete proof in progress, err: %v", err2)
+			}
+		}
+		log.Debug("tryGenerateBatchProof end")
+	}()
+
+	log.Info("Generating proof from batch")
+
+	log.Infof("Sending zki + batch to the prover, batchNumber [%d]", batchToProve.BatchNumber)
+	inputProver, err := a.buildInputProver(ctx, batchToProve)
+	if err != nil {
+		err = fmt.Errorf("failed to build input prover, %w", err)
+		log.Error(FirstToUpper(err.Error()))
+		return false, err
+	}
+
+	b, err := json.Marshal(inputProver)
+	if err != nil {
+		err = fmt.Errorf("failed to serialize input prover, %w", err)
+		log.Error(FirstToUpper(err.Error()))
+		return false, err
+	}
+
+	proof.InputProver = string(b)
+
+	log.Infof("Sending a batch to the prover. OldStateRoot [%#x], OldBatchNum [%d]",
+		inputProver.PublicInputs.OldStateRoot, inputProver.PublicInputs.OldBatchNum)
+
+	genProofID, err = prover.BatchProof(inputProver)
+	if err != nil {
+		err = fmt.Errorf("failed to get batch proof id, %w", err)
+		log.Error(FirstToUpper(err.Error()))
+		return false, err
+	}
+
+	proof.ProofID = genProofID
+
+	log.Infof("Proof ID %v", *proof.ProofID)
+	log = log.WithFields("proofId", *proof.ProofID)
+
+	resGetProof, err := prover.WaitRecursiveProof(ctx, *proof.ProofID)
+	if err != nil {
+		err = fmt.Errorf("failed to get proof from prover, %w", err)
+		log.Error(FirstToUpper(err.Error()))
+		return false, err
+	}
+
+	log.Info("Batch proof generated")
+
+	proof.Proof = resGetProof
+
+	// NOTE(pg): the defer func is useless from now on, use a different variable
+	// name for errors (or shadow err in inner scopes) to not trigger it.
+
+	finalProofBuilt, finalProofErr := a.tryBuildFinalProof(ctx, prover, proof)
+	if finalProofErr != nil {
+		// just log the error and continue to handle the generated proof
+		log.Errorf("Error trying to build final proof: %v", finalProofErr)
+	}
+
+	// NOTE(pg): prover is done, use a.ctx from now on
+
+	if !finalProofBuilt {
+		proof.GeneratingSince = nil
+
+		// final proof has not been generated, update the batch proof
+		err := a.State.UpdateBatchProof(a.ctx, proof, nil)
+		if err != nil {
+			err = fmt.Errorf("failed to store batch proof result, %w", err)
+			log.Error(FirstToUpper(err.Error()))
+			return false, err
+		}
+	}
+
+	return true, nil
+}
+
+func (a *Aggregator) getAndLockBatchToProve(ctx context.Context, prover proverInterface) (*state.Batch, *state.Proof, error) {
+	proverID := prover.ID()
+	proverName := prover.Name()
+
+	log := log.WithFields(
+		"prover", proverName,
+		"proverId", proverID,
+		"proverAddr", prover.Addr(),
+	)
+
+	a.StateDBMutex.Lock()
+	defer a.StateDBMutex.Unlock()
+
+	lastVerifiedBatch, err := a.State.GetLastVerifiedBatch(ctx, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	// Get header of the last L1 block
+	lastL1BlockHeader, err := a.Ethman.GetLatestBlockHeader(ctx)
+	if err != nil {
+		log.Errorf("Failed to get last L1 block header, err: %v", err)
+		return nil, nil, err
+	}
+	lastL1BlockNumber := lastL1BlockHeader.Number.Uint64()
+
+	// Calculate max L1 block number for getting next virtual batch to prove
+	maxL1BlockNumber := uint64(0)
+	if a.cfg.BatchProofL1BlockConfirmations <= lastL1BlockNumber {
+		maxL1BlockNumber = lastL1BlockNumber - a.cfg.BatchProofL1BlockConfirmations
+	}
+	log.Debugf("Max L1 block number for getting next virtual batch to prove: %d", maxL1BlockNumber)
+
+	// Get virtual batch pending to generate proof
+	batchToVerify, err := a.State.GetVirtualBatchToProve(ctx, lastVerifiedBatch.BatchNumber, maxL1BlockNumber, nil)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	log.Infof("Found virtual batch %d pending to generate proof", batchToVerify.BatchNumber)
+	log = log.WithFields("batch", batchToVerify.BatchNumber)
+
+	log.Info("Checking profitability to aggregate batch")
+
+	// pass pol collateral as zero here, bcs in smart contract fee for aggregator is not defined yet
+	isProfitable, err := a.ProfitabilityChecker.IsProfitable(ctx, big.NewInt(0))
+	if err != nil {
+		log.Errorf("Failed to check aggregator profitability, err: %v", err)
+		return nil, nil, err
+	}
+
+	if !isProfitable {
+		log.Infof("Batch is not profitable, pol collateral %d", big.NewInt(0))
+		return nil, nil, err
+	}
+
+	now := time.Now().Round(time.Microsecond)
+	proof := &state.Proof{
+		BatchNumber:      batchToVerify.BatchNumber,
+		BatchNumberFinal: batchToVerify.BatchNumber,
+		Prover:           &proverName,
+		ProverID:         &proverID,
+		GeneratingSince:  &now,
+	}
+
+	// Avoid other prover to process the same batch
+	err = a.State.AddBatchProof(ctx, proof, nil)
+	if err != nil {
+		log.Errorf("Failed to add batch proof, err: %v", err)
+		return nil, nil, err
+	}
+
+	return batchToVerify, proof, nil
+}
