+++ aggregator/blobinner.go
@@ -1,0 +1,7 @@
+package aggregator
+
+import "context"
+
+func (a *Aggregator) tryGenerateBlobInnerProof(ctx context.Context, prover proverInterface) (bool, error) {
+	return false, nil
+}
