+++ state/pgstatestorage/proof.go
@@ -25,9 +25,9 @@ func (p *PostgresStorage) CheckProofContainsCompleteSequences(ctx context.Contex
 	return exists, nil
 }
 
-// GetProofReadyToVerify return the proof that is ready to verify
-func (p *PostgresStorage) GetProofReadyToVerify(ctx context.Context, lastVerfiedBatchNumber uint64, dbTx pgx.Tx) (*state.Proof, error) {
-	const getProofReadyToVerifySQL = `
+// GetProofReadyForFinal return the proof that is ready to generate the final proof
+func (p *PostgresStorage) GetProofReadyForFinal(ctx context.Context, lastVerfiedBatchNumber uint64, dbTx pgx.Tx) (*state.Proof, error) {
+	const getProofReadyForFinalSQL = `
 		SELECT 
 			p.batch_num, 
 			p.batch_num_final,
@@ -48,7 +48,7 @@ func (p *PostgresStorage) GetProofReadyToVerify(ctx context.Context, lastVerfied
 	var proof *state.Proof = &state.Proof{}
 
 	e := p.getExecQuerier(dbTx)
-	row := e.QueryRow(ctx, getProofReadyToVerifySQL, lastVerfiedBatchNumber+1)
+	row := e.QueryRow(ctx, getProofReadyForFinalSQL, lastVerfiedBatchNumber+1)
 	err := row.Scan(&proof.BatchNumber, &proof.BatchNumberFinal, &proof.Proof, &proof.ProofID, &proof.InputProver, &proof.Prover, &proof.ProverID, &proof.GeneratingSince, &proof.CreatedAt, &proof.UpdatedAt)
 
 	if errors.Is(err, pgx.ErrNoRows) {
@@ -60,15 +60,15 @@ func (p *PostgresStorage) GetProofReadyToVerify(ctx context.Context, lastVerfied
 	return proof, err
 }
 
-// GetProofsToAggregate return the next to proof that it is possible to aggregate
-func (p *PostgresStorage) GetProofsToAggregate(ctx context.Context, dbTx pgx.Tx) (*state.Proof, *state.Proof, error) {
+// GetBatchProofsToAggregate return the next 2 batch proofs that it are possible to aggregate
+func (p *PostgresStorage) GetBatchProofsToAggregate(ctx context.Context, dbTx pgx.Tx) (*state.Proof, *state.Proof, error) {
 	var (
 		proof1 *state.Proof = &state.Proof{}
 		proof2 *state.Proof = &state.Proof{}
 	)
 
 	// TODO: add comments to explain the query
-	const getProofsToAggregateSQL = `
+	const getBatchProofsToAggregateSQL = `
 		SELECT 
 			p1.batch_num as p1_batch_num, 
 			p1.batch_num_final as p1_batch_num_final, 
@@ -91,30 +91,15 @@ func (p *PostgresStorage) GetProofsToAggregate(ctx context.Context, dbTx pgx.Tx)
 			p2.created_at as p2_created_at,
 			p2.updated_at as p2_updated_at
 		FROM state.proof p1 INNER JOIN state.proof p2 ON p1.batch_num_final = p2.batch_num - 1
-		WHERE p1.generating_since IS NULL AND p2.generating_since IS NULL AND 
+		WHERE p1.blob_inner_num = p2.blob_inner_num AND
+			  p1.generating_since IS NULL AND p2.generating_since IS NULL AND 
 		 	  p1.proof IS NOT NULL AND p2.proof IS NOT NULL AND
-			  (
-					EXISTS (
-					SELECT 1 FROM state.sequences s
-					WHERE p1.batch_num >= s.from_batch_num AND p1.batch_num <= s.to_batch_num AND
-						p1.batch_num_final >= s.from_batch_num AND p1.batch_num_final <= s.to_batch_num AND
-						p2.batch_num >= s.from_batch_num AND p2.batch_num <= s.to_batch_num AND
-						p2.batch_num_final >= s.from_batch_num AND p2.batch_num_final <= s.to_batch_num
-					)
-					OR
-					(
-						EXISTS ( SELECT 1 FROM state.sequences s WHERE p1.batch_num = s.from_batch_num) AND
-						EXISTS ( SELECT 1 FROM state.sequences s WHERE p1.batch_num_final = s.to_batch_num) AND
-						EXISTS ( SELECT 1 FROM state.sequences s WHERE p2.batch_num = s.from_batch_num) AND
-						EXISTS ( SELECT 1 FROM state.sequences s WHERE p2.batch_num_final = s.to_batch_num)
-					)
-				)
 		ORDER BY p1.batch_num ASC
 		LIMIT 1
 		`
 
 	e := p.getExecQuerier(dbTx)
-	row := e.QueryRow(ctx, getProofsToAggregateSQL)
+	row := e.QueryRow(ctx, getBatchProofsToAggregateSQL)
 	err := row.Scan(
 		&proof1.BatchNumber, &proof1.BatchNumberFinal, &proof1.Proof, &proof1.ProofID, &proof1.InputProver, &proof1.Prover, &proof1.ProverID, &proof1.GeneratingSince, &proof1.CreatedAt, &proof1.UpdatedAt,
 		&proof2.BatchNumber, &proof2.BatchNumberFinal, &proof2.Proof, &proof2.ProofID, &proof2.InputProver, &proof2.Prover, &proof2.ProverID, &proof2.GeneratingSince, &proof2.CreatedAt, &proof2.UpdatedAt)
@@ -128,45 +113,42 @@ func (p *PostgresStorage) GetProofsToAggregate(ctx context.Context, dbTx pgx.Tx)
 	return proof1, proof2, err
 }
 
-// AddGeneratedProof adds a generated proof to the storage
-func (p *PostgresStorage) AddGeneratedProof(ctx context.Context, proof *state.Proof, dbTx pgx.Tx) error {
-	const addGeneratedProofSQL = "INSERT INTO state.proof (batch_num, batch_num_final, proof, proof_id, input_prover, prover, prover_id, generating_since, created_at, updated_at) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)"
+// AddBatchProof adds a batch proof to the storage
+func (p *PostgresStorage) AddBatchProof(ctx context.Context, proof *state.Proof, dbTx pgx.Tx) error {
+	const addBatchProofSQL = "INSERT INTO state.proof (batch_num, batch_num_final, proof, proof_id, input_prover, prover, prover_id, generating_since, created_at, updated_at) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)"
 	e := p.getExecQuerier(dbTx)
 	now := time.Now().UTC().Round(time.Microsecond)
-	_, err := e.Exec(ctx, addGeneratedProofSQL, proof.BatchNumber, proof.BatchNumberFinal, proof.Proof, proof.ProofID, proof.InputProver, proof.Prover, proof.ProverID, proof.GeneratingSince, now, now)
+	_, err := e.Exec(ctx, addBatchProofSQL, proof.BatchNumber, proof.BatchNumberFinal, proof.Proof, proof.ProofID, proof.InputProver, proof.Prover, proof.ProverID, proof.GeneratingSince, now, now)
 	return err
 }
 
-// UpdateGeneratedProof updates a generated proof in the storage
-func (p *PostgresStorage) UpdateGeneratedProof(ctx context.Context, proof *state.Proof, dbTx pgx.Tx) error {
-	const addGeneratedProofSQL = "UPDATE state.proof SET proof = $3, proof_id = $4, input_prover = $5, prover = $6, prover_id = $7, generating_since = $8, updated_at = $9 WHERE batch_num = $1 AND batch_num_final = $2"
+// UpdateBatchProof updates a batch proof in the storage
+func (p *PostgresStorage) UpdateBatchProof(ctx context.Context, proof *state.Proof, dbTx pgx.Tx) error {
+	const addBatchProofSQL = "UPDATE state.proof SET proof = $3, proof_id = $4, input_prover = $5, prover = $6, prover_id = $7, generating_since = $8, updated_at = $9 WHERE batch_num = $1 AND batch_num_final = $2"
 	e := p.getExecQuerier(dbTx)
 	now := time.Now().UTC().Round(time.Microsecond)
-	_, err := e.Exec(ctx, addGeneratedProofSQL, proof.BatchNumber, proof.BatchNumberFinal, proof.Proof, proof.ProofID, proof.InputProver, proof.Prover, proof.ProverID, proof.GeneratingSince, now)
+	_, err := e.Exec(ctx, addBatchProofSQL, proof.BatchNumber, proof.BatchNumberFinal, proof.Proof, proof.ProofID, proof.InputProver, proof.Prover, proof.ProverID, proof.GeneratingSince, now)
 	return err
 }
 
-// DeleteGeneratedProofs deletes from the storage the generated proofs falling
-// inside the batch numbers range.
-func (p *PostgresStorage) DeleteGeneratedProofs(ctx context.Context, batchNumber uint64, batchNumberFinal uint64, dbTx pgx.Tx) error {
-	const deleteGeneratedProofSQL = "DELETE FROM state.proof WHERE batch_num >= $1 AND batch_num_final <= $2"
+// DeleteBatchProofs deletes from the storage the batch proofs falling inside the batch numbers range.
+func (p *PostgresStorage) DeleteBatchProofs(ctx context.Context, batchNumber uint64, batchNumberFinal uint64, dbTx pgx.Tx) error {
+	const deleteBatchProofSQL = "DELETE FROM state.proof WHERE batch_num >= $1 AND batch_num_final <= $2"
 	e := p.getExecQuerier(dbTx)
-	_, err := e.Exec(ctx, deleteGeneratedProofSQL, batchNumber, batchNumberFinal)
+	_, err := e.Exec(ctx, deleteBatchProofSQL, batchNumber, batchNumberFinal)
 	return err
 }
 
-// CleanupGeneratedProofs deletes from the storage the generated proofs up to
-// the specified batch number included.
-func (p *PostgresStorage) CleanupGeneratedProofs(ctx context.Context, batchNumber uint64, dbTx pgx.Tx) error {
-	const deleteGeneratedProofSQL = "DELETE FROM state.proof WHERE batch_num_final <= $1"
+// CleanupBatchProofs deletes from the storage the batch proofs up to the specified batch number included.
+func (p *PostgresStorage) CleanupBatchProofs(ctx context.Context, batchNumber uint64, dbTx pgx.Tx) error {
+	const deleteBatchProofSQL = "DELETE FROM state.proof WHERE batch_num_final <= $1"
 	e := p.getExecQuerier(dbTx)
-	_, err := e.Exec(ctx, deleteGeneratedProofSQL, batchNumber)
+	_, err := e.Exec(ctx, deleteBatchProofSQL, batchNumber)
 	return err
 }
 
-// CleanupLockedProofs deletes from the storage the proofs locked in generating
-// state for more than the provided threshold.
-func (p *PostgresStorage) CleanupLockedProofs(ctx context.Context, duration string, dbTx pgx.Tx) (int64, error) {
+// CleanupLockedBatchProofs deletes from the storage the proofs locked in generating state for more than the provided threshold.
+func (p *PostgresStorage) CleanupLockedBatchProofs(ctx context.Context, duration string, dbTx pgx.Tx) (int64, error) {
 	interval, err := toPostgresInterval(duration)
 	if err != nil {
 		return 0, err
@@ -180,9 +162,8 @@ func (p *PostgresStorage) CleanupLockedProofs(ctx context.Context, duration stri
 	return ct.RowsAffected(), nil
 }
 
-// DeleteUngeneratedProofs deletes ungenerated proofs.
-// This method is meant to be use during aggregator boot-up sequence
-func (p *PostgresStorage) DeleteUngeneratedProofs(ctx context.Context, dbTx pgx.Tx) error {
+// DeleteUngeneratedBatchProofs deletes ungenerated proofs. This method is meant to be use during aggregator boot-up sequence
+func (p *PostgresStorage) DeleteUngeneratedBatchProofs(ctx context.Context, dbTx pgx.Tx) error {
 	const deleteUngeneratedProofsSQL = "DELETE FROM state.proof WHERE generating_since IS NOT NULL"
 	e := p.getExecQuerier(dbTx)
 	_, err := e.Exec(ctx, deleteUngeneratedProofsSQL)
